More changes to make the store work properly

+++ store/sqlstore/supplier.go
@@ -11,6 +11,7 @@ import (
 	"fmt"
 	sqltrace "log"
 	"os"
+	"regexp"
 	"strings"
 	"sync/atomic"
 	"time"
@@ -567,6 +568,25 @@ func (ss *SqlSupplier) DoesTriggerExist(triggerName string) bool {
 
 		return count > 0
 
+	} else if ss.DriverName() == model.DATABASE_DRIVER_SQLITE {
+		count, err := ss.GetMaster().SelectInt(`
+			SELECT
+		    	COUNT(0)
+			FROM
+			    sqlite_master
+			WHERE
+			    type='trigger'
+			    AND name = ?
+		`, triggerName)
+
+		if err != nil {
+			mlog.Critical("Failed to check if trigger exists", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_GENERIC_FAILURE)
+		}
+
+		return count > 0
+
 	} else {
 		mlog.Critical("Failed to check if column exists because of missing driver")
 		time.Sleep(time.Second)
@@ -600,7 +620,15 @@ func (ss *SqlSupplier) CreateColumnIfNotExists(tableName string, columnName stri
 		}
 
 		return true
+	} else if ss.DriverName() == model.DATABASE_DRIVER_SQLITE {
+		_, err := ss.GetMaster().ExecNoTimeout("ALTER TABLE " + tableName + " ADD " + columnName + " " + mySqlColType + " DEFAULT '" + defaultValue + "'")
+		if err != nil {
+			mlog.Critical("Failed to create column", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_CREATE_COLUMN_MYSQL)
+		}
 
+		return true
 	} else {
 		mlog.Critical("Failed to create column because of missing driver")
 		time.Sleep(time.Second)
@@ -657,6 +685,10 @@ func (ss *SqlSupplier) RemoveColumnIfExists(tableName string, columnName string)
 
 	if !ss.DoesColumnExist(tableName, columnName) {
 		return false
+	}
+
+	if ss.DriverName() == model.DATABASE_DRIVER_SQLITE {
+		return true
 	}
 
 	_, err := ss.GetMaster().ExecNoTimeout("ALTER TABLE " + tableName + " DROP COLUMN " + columnName)
@@ -717,7 +749,7 @@ func (ss *SqlSupplier) GetMaxLengthOfColumnIfExists(tableName string, columnName
 	if ss.DriverName() == model.DATABASE_DRIVER_MYSQL {
 		result, err = ss.GetMaster().SelectStr("SELECT CHARACTER_MAXIMUM_LENGTH FROM information_schema.columns WHERE table_name = '" + tableName + "' AND COLUMN_NAME = '" + columnName + "'")
 	} else if ss.DriverName() == model.DATABASE_DRIVER_SQLITE {
-		result, err = ss.GetMaster().SelectStr("SELECT CHARACTER_MAXIMUM_LENGTH FROM information_schema.columns WHERE table_name = '" + tableName + "' AND COLUMN_NAME = '" + columnName + "'")
+		result = "2147483647"
 	} else if ss.DriverName() == model.DATABASE_DRIVER_POSTGRES {
 		result, err = ss.GetMaster().SelectStr("SELECT character_maximum_length FROM information_schema.columns WHERE table_name = '" + strings.ToLower(tableName) + "' AND column_name = '" + strings.ToLower(columnName) + "'")
 	}
@@ -740,7 +772,38 @@ func (ss *SqlSupplier) AlterColumnTypeIfExists(tableName string, columnName stri
 	if ss.DriverName() == model.DATABASE_DRIVER_MYSQL {
 		_, err = ss.GetMaster().ExecNoTimeout("ALTER TABLE " + tableName + " MODIFY " + columnName + " " + mySqlColType)
 	} else if ss.DriverName() == model.DATABASE_DRIVER_SQLITE {
-		_, err = ss.GetMaster().ExecNoTimeout("ALTER TABLE " + tableName + " MODIFY " + columnName + " " + mySqlColType)
+		createTableQuery, err := ss.GetMaster().SelectStr(`SELECT sql from sqlite_master where type="table" AND name="` + tableName + `"`)
+		if err != nil {
+			mlog.Critical("Failed to alter column type", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_ALTER_COLUMN)
+		}
+		_, err = ss.GetMaster().ExecNoTimeout("ALTER TABLE " + tableName + " RENAME TO tmp")
+		if err != nil {
+			mlog.Critical("Failed to alter column type", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_ALTER_COLUMN)
+		}
+		re := regexp.MustCompile(columnName + " ([^,) ])")
+		createTableQuery = re.ReplaceAllString(createTableQuery, columnName+" "+mySqlColType+"$1")
+		_, err = ss.GetMaster().ExecNoTimeout(createTableQuery)
+		if err != nil {
+			mlog.Critical("Failed to alter column type", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_ALTER_COLUMN)
+		}
+		_, err = ss.GetMaster().ExecNoTimeout("INSERT INTO " + tableName + " SELECT * from tmp")
+		if err != nil {
+			mlog.Critical("Failed to alter column type", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_ALTER_COLUMN)
+		}
+		_, err = ss.GetMaster().ExecNoTimeout("DROP TABLE tmp")
+		if err != nil {
+			mlog.Critical("Failed to alter column type", mlog.Err(err))
+			time.Sleep(time.Second)
+			os.Exit(EXIT_ALTER_COLUMN)
+		}
 	} else if ss.DriverName() == model.DATABASE_DRIVER_POSTGRES {
 		_, err = ss.GetMaster().ExecNoTimeout("ALTER TABLE " + strings.ToLower(tableName) + " ALTER COLUMN " + strings.ToLower(columnName) + " TYPE " + postgresColType)
 	}
